decouple processing pipelines
