+++ Smartscope/core/pipelines/__init__.py
@@ -1,0 +1,3 @@
+
+
+from .preprocessing_pipeline_cmf import PreprocessingPipelineCmd
