+++ Smartscope/core/settings/local.py
@@ -107,7 +107,7 @@ CHANNEL_LAYERS = {
     "default": {
         "BACKEND": "channels_redis.core.RedisChannelLayer",
         "CONFIG": {
-            "hosts": [REDIS_HOST, REDIS_PORT],
+            "hosts": [(REDIS_HOST, REDIS_PORT)],
         },
     },
 }
